# api/openapi/main.yaml
openapi: 3.0.3
info:
  title: OT Continuum API
  version: 1.0.0
  description: |
    Multi-tenant operational technology risk management platform.
    
    ## Authentication
    All endpoints require a Bearer token in the Authorization header:
    ```
    Authorization: Bearer <your_token>
    ```
    
    ## Multi-Tenancy
    All requests are automatically scoped to the authenticated user's tenant via JWT claims.
    
    ## Rate Limiting
    - 1000 requests per hour per user
    - 10000 requests per hour per tenant
  contact:
    name: OT Continuum API Support
    email: api-support@otcontinuum.com
  license:
    name: Proprietary
    url: https://otcontinuum.com/license
servers:
  - url: https://{project_id}.supabase.co/functions/v1/make-server-fb677d93
    description: Production server
    variables:
      project_id:
        default: your-project-id
        description: Your Supabase project ID
  - url: http://localhost:54321/functions/v1/make-server-fb677d93
    description: Local development server
...
security:
  - BearerAuth: []
paths:
...
  /users:
    get:
      tags: [users]
      summary: List users in tenant
      description: Returns all users in the authenticated user's tenant
      operationId: listUsers
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: role
          in: query
          schema:
            $ref: '#/components/schemas/UserRole'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/UserStatus'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags: [users]
      summary: Create new user
      description: Create a new user in the tenant (admin only)
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /users/{userId}:
    get:
      tags: [users]
      summary: Get user by ID
      operationId: getUser
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    patch:
      tags: [users]
      summary: Update user
      description: Update user profile (admin or self)
      operationId: updateUser
      parameters:
        - $ref: '#/components/parameters/UserId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'

    delete:
      tags: [users]
      summary: Delete user
      description: Soft delete user (admin only)
      operationId: deleteUser
      parameters:
        - $ref: '#/components/parameters/UserId'
      responses:
        '204':
          description: User deleted successfully
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /risks:
    get:
      tags: [risks]
      summary: List risks
      description: Returns risks accessible to the user (with ownership bypass)
      operationId: listRisks
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
        - name: site_id
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/RiskStatus'
        - name: decision
          in: query
          schema:
            $ref: '#/components/schemas/RiskDecision'
        - name: owner_id
          in: query
          schema:
            type: string
            format: uuid
        - name: min_risk_score
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 25
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Risk'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags: [risks]
      summary: Create risk
      description: Create a new risk (contributor or higher)
      operationId: createRisk
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RiskCreate'
      responses:
        '201':
          description: Risk created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Risk'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /risks/{riskId}:
    get:
      tags: [risks]
      summary: Get risk by ID
      operationId: getRisk
      parameters:
        - $ref: '#/components/parameters/RiskId'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Risk'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

    patch:
      tags: [risks]
      summary: Update risk
      description: Update risk (owner can always update their risks)
      operationId: updateRisk
      parameters:
        - $ref: '#/components/parameters/RiskId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RiskUpdate'
      responses:
        '200':
          description: Risk updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Risk'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from Supabase Auth
  parameters:
    Limit:
      name: limit
      in: query
      description: Maximum number of results to return
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      description: Number of results to skip
      schema:
        type: integer
        minimum: 0
        default: 0
    UserId:
      name: userId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    RiskId:
      name: riskId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    Severity:
      type: string
      enum: [catastrophic, major, moderate, minor, negligible]
    Likelihood:
      type: string
      enum: [almost_certain, likely, possible, unlikely, rare]
    RiskDecision:
      type: string
      enum: [accept, mitigate, transfer, avoid, under_review]
    RiskStatus:
      type: string
      enum: [open, monitoring, closed, escalated]
...
    UserRole:
      type: string
      enum: [admin, manager, operator, viewer, auditor, contributor, billing_admin]
    UserStatus:
      type: string
      enum: [active, inactive, suspended, pending]
...
    Pagination:
      type: object
      properties:
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer
    Timestamps:
      type: object
      properties:
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
    ValidationError:
      allOf:
        - $ref: '#/components/schemas/Error'
        - type: object
          properties:
            errors:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string

    # =============================================================================
    # TENANT
    # =============================================================================
...
    User:
      allOf:
        - $ref: '#/components/schemas/Timestamps'
        - type: object
          required:
            - id
            - tenant_id
            - email
            - role
            - status
          properties:
            id:
              type: string
              format: uuid
            tenant_id:
              type: string
              format: uuid
            email:
              type: string
              format: email
            full_name:
              type: string
              nullable: true
            avatar_url:
              type: string
              nullable: true
            role:
              $ref: '#/components/schemas/UserRole'
            status:
              $ref: '#/components/schemas/UserStatus'
            site_ids:
              type: array
              items:
                type: string
                format: uuid
              nullable: true
            last_seen_at:
              type: string
              format: date-time
              nullable: true
    UserCreate:
      type: object
      required:
        - email
        - role
      properties:
        email:
          type: string
          format: email
        full_name:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        site_ids:
          type: array
          items:
            type: string
            format: uuid
    UserUpdate:
      type: object
      properties:
        full_name:
          type: string
        avatar_url:
          type: string
        role:
          $ref: '#/components/schemas/UserRole'
        status:
          $ref: '#/components/schemas/UserStatus'
        site_ids:
          type: array
          items:
            type: string
            format: uuid

    # =============================================================================
    # SITE
    # =============================================================================
...
    Risk:
      allOf:
        - $ref: '#/components/schemas/Timestamps'
        - type: object
          required:
            - id
            - tenant_id
            - site_id
            - risk_id
            - title
            - description
            - severity
            - likelihood
            - risk_score
            - owner_id
            - decision
            - status
          properties:
            id:
              type: string
              format: uuid
            tenant_id:
              type: string
              format: uuid
            site_id:
              type: string
              format: uuid
            risk_id:
              type: string
            title:
              type: string
            description:
              type: string
            category:
              type: string
              nullable: true
            severity:
              $ref: '#/components/schemas/Severity'
            likelihood:
              $ref: '#/components/schemas/Likelihood'
            risk_score:
              type: integer
              readOnly: true
            owner_id:
              type: string
              format: uuid
            decision:
              $ref: '#/components/schemas/RiskDecision'
            decision_rationale:
              type: string
              nullable: true
            decision_date:
              type: string
              format: date-time
              nullable: true
            decision_by:
              type: string
              format: uuid
              nullable: true
            status:
              $ref: '#/components/schemas/RiskStatus'
            existing_controls:
              type: string
              nullable: true
            mitigation_plan:
              type: string
              nullable: true
            target_completion_date:
              type: string
              format: date-time
              nullable: true
            related_work_item_id:
              type: string
              format: uuid
              nullable: true
            audit_trail:
              type: array
              items:
                type: object
            last_reviewed_at:
              type: string
              format: date-time
              nullable: true
            next_review_due:
              type: string
              format: date-time
              nullable: true
            closed_at:
              type: string
              format: date-time
              nullable: true
    RiskCreate:
      type: object
      required:
        - site_id
        - risk_id
        - title
        - description
        - severity
        - likelihood
        - owner_id
      properties:
        site_id:
          type: string
          format: uuid
        risk_id:
          type: string
        title:
          type: string
        description:
          type: string
        category:
          type: string
        severity:
          $ref: '#/components/schemas/Severity'
        likelihood:
          $ref: '#/components/schemas/Likelihood'
        owner_id:
          type: string
          format: uuid
        decision:
          $ref: '#/components/schemas/RiskDecision'
        decision_rationale:
          type: string
        existing_controls:
          type: string
        mitigation_plan:
          type: string
        target_completion_date:
          type: string
          format: date-time
    RiskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        category:
          type: string
        severity:
          $ref: '#/components/schemas/Severity'
        likelihood:
          $ref: '#/components/schemas/Likelihood'
        owner_id:
          type: string
          format: uuid
        decision:
          $ref: '#/components/schemas/RiskDecision'
        decision_rationale:
          type: string
        status:
          $ref: '#/components/schemas/RiskStatus'
        existing_controls:
          type: string
        mitigation_plan:
          type: string
        target_completion_date:
          type: string
          format: date-time
        next_review_due:
          type: string
          format: date-time

    # =============================================================================
    # RISK EVENT
    # =============================================================================
...
  responses:
    Unauthorized:
      description: Unauthorized - Invalid or missing authentication token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: unauthorized
            message: Authentication token is required
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: forbidden
            message: You do not have permission to perform this action
    NotFound:
      description: Not found - Resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: not_found
            message: The requested resource was not found
    ValidationError:
      description: Validation error - Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
          example:
            error: validation_error
            message: Request validation failed
            errors:
              - field: email
                message: Invalid email format
              - field: site_id
                message: Site ID is required
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: internal_error
            message: An unexpected error occurred